requirements.txt
yt-dlp
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchaudio
pyannote.audio
demucs
# jupyter-lab
# tensorboard
